server/requirements.txt
# Planting proof API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.35.0
slowapi>=0.1.9
httpx>=0.28.0
pillow>=10.4.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
ultralytics>=8.3.0
# torch / torchvision: install separately (e.g. CPU wheels in Docker) so CUDA meta-packages are not pulled in.
